--- addons/l10n_in_withholding_payment/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import wizard
