--- addons/l10n_in_withholding_payment/__manifest__.py
@@ -1,11 +1,0 @@
-{
-    'name': 'Indian - TDS For Payment',
-    'version': '1.0',
-    'description': """
-        Support for Indian TDS (Tax Deducted at Source) for Payment.
-    """,
-    'category': 'Accounting/Localizations',
-    'depends': ['l10n_in_withholding'],
-    'auto_install': True,
-    'license': 'LGPL-3',
-}
