--- addons/l10n_in_withholding_payment/models/__init__.py
@@ -1,2 +1,0 @@
-from . import account_move
-from . import account_payment
