--- addons/l10n_in_withholding_payment/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import l10n_in_withhold_wizard
